[disk] AgentCheck for Disk metrics :tada:
This commit is the first step towards the annihilation of the old `Check`
hidden in `checks/system/{unix,win32}.py`!
It's time to launch the final assault of `Check`. `AgentCheck` shall triumph!

Add a `checks.d/disk.py` checks, which uses psutil when available to get
disk metrics (same than the previous `disk` checks), and otherwise
get the same metrics with `df`.
It has a small integration test, just to check that metrics exist, and
more precise mock tests.
This new check is enabled by default, and the old `disk` checks are
disabled.