[install script] Add DD_NO_START option
This new option can be used in the one-liner script to prevent the agent
from starting automatically at the end of the installation. It is useful
to run the script with this option to create prebaked images where you
want to start the agent only at a later stage when deploying.

[skip ci]